Eliminate mypyc.buildc
